fix: Resolve all 73 mypy type errors for full type safety
- Added missing type parameters for generic types (list, dict, ndarray, Callable)
- Added type: ignore comments for missing library stubs (tensorflow, scipy, joblib, pyarrow)
- Removed unused type: ignore comments
- Fixed import issues and type mismatches
- Fixed hypothesis test health check warnings
- Achieved 0 mypy errors across all 165 source files

+++ src/big_mood_detector/domain/services/pat_sequence_builder.py
@@ -39,7 +39,7 @@ class PATSequence:
     """
 
     end_date: date  # Last day of the sequence
-    activity_values: np.ndarray[Any, np.dtype[np.float64]]  # Shape: (10080,)
+    activity_values: np.ndarray[Any, np.dtype[np.float32]]  # Shape: (10080,)
     missing_days: list[date]  # Days with no data
     data_quality_score: float  # 0-1, based on completeness
 
@@ -59,7 +59,7 @@ class PATSequence:
         """Check if all 7 days have data."""
         return len(self.missing_days) == 0
 
-    def to_patches(self, patch_size: int = 18) -> np.ndarray[Any, np.dtype[np.float64]]:
+    def to_patches(self, patch_size: int = 18) -> np.ndarray[Any, np.dtype[np.float32]]:
         """
         Convert to patches for transformer input.
 
@@ -74,7 +74,7 @@ class PATSequence:
             num_patches, patch_size
         )
 
-    def get_normalized(self) -> np.ndarray[Any, np.dtype[np.float64]]:
+    def get_normalized(self) -> np.ndarray[Any, np.dtype[np.float32]]:
         """
         Get z-score normalized sequence.
 
@@ -206,7 +206,7 @@ class PATSequenceBuilder:
 
     def _combine_sequences(
         self, daily_sequences: list[MinuteLevelSequence]
-    ) -> np.ndarray[Any, np.dtype[np.float64]]:
+    ) -> np.ndarray[Any, np.dtype[np.float32]]:
         """
         Combine daily sequences into a single 7-day array.
 
@@ -225,7 +225,7 @@ class PATSequenceBuilder:
 
     def _interpolate_missing_days(
         self, values: NDArray[np.float32], missing_days: list[date], start_date: date
-    ) -> np.ndarray[Any, np.dtype[np.float64]]:
+    ) -> np.ndarray[Any, np.dtype[np.float32]]:
         """
         Interpolate missing days using neighboring data.
